requirements.txt
fastapi
uvicorn
groq
langgraph
langchain-core
langchain-mcp-adapters
fastmcp
httpx
python-dotenv
pydantic